Move experiment qualification to after waiting room
If an experiment is overrecruiting then the experiment qualification
should only be given to users who are actually shown the experiment.
If someone joins then is discarded by the waiting room they should
be unqualified. Assigning then removing the qualification causes
spurious emails to be generated by MTurk, so we change the way
overrecruitment is defined to be a python function rather than
comparing two elements of a JSON value in Javascript. This lets
us assign the qualification in Python code and makes for a slightly
more standardised override story for experimenters.